Rename FinalMenu to Menu and put into own file.
